requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
llama-cpp-python
numpy
pymupdf